Split up Limit
